Renamed files according to the new name
