+++ lib/mockingbird.js
